drivers: clk: workaround si5338 initialization bug
One of the clocks in the si5338 clock tree has rate exceeding maximum
value possible for long type. This was causing the clk_calc_new_rates
function to fail as the frequency value was corrupted. Changing the type
from long to unsigned long fixed the issue. It shouldn't influence the
overall clk framework as the frequency cannot be negative.