Add support for Mercury XU61
Due to changes to the MIO mapping of the XU61, changes in all other Enclustra Mercury and Mars XU* device tree files are necessary:
- moving all i2c entries to the module common dtsi files since the XU61 uses the i2c1 PS controller instead of i2c0 as all other previous XU* modules

Furthermore, changes in the buildscripts repository led to a different order in which the enclustra_generated.dts is including the various .dtsi files. Therefore, the includes have been removed from module specific .dtsi files.